pyproject.toml
[project]
name = "detection-notes"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "ipykernel>=6.30.1",
    "mistralai>=1.9.10",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.1.1",
    "opencv-python>=4.10.0",
    "numpy>=2.2.6",
    "streamlit>=1.50.0",
    "paramiko>=4.0.0",
    "scp>=0.15.0",
    "openai-whisper",
    "paramiko>=4.0.0",
    "scp>=0.15.0",
    "ultralytics>=8.3.203",
    "lap>=0.5.12",
    "pytubefix>=9.5.1",
    "scikit-image>=0.25.2",
    "watchdog>=6.0.0",
    "transformers>=4.56.2",
    "pot>=0.9.6.post1",
    "datasets>=4.1.1",
    "jiwer>=4.0.0",
    "gradio>=5.48.0",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "seqeval>=1.2.2",
    "torch>=2.8.0",
    "pyzbar>=0.1.9",
    "noisereduce>=3.0.3",
    "torchaudio>=2.8.0",
    "huggingface-hub>=0.35.3",
    "pyannote-audio>=4.0.1",
    "librosa>=0.11.0",
]
